+++ packages/react-learn-card/src/components/CertificateDisplayCard/CertificateFrontFace.tsx
@@ -36,6 +36,7 @@ type CertificateFrontFaceProps = {
     hideIssueDate?: boolean;
     handleViewBackFace?: () => void;
     showDetailsBtn?: boolean;
+    formattedDisplayType?: string;
 };
 
 export const CertificateFrontFace: React.FC<CertificateFrontFaceProps> = ({
@@ -51,6 +52,7 @@ export const CertificateFrontFace: React.FC<CertificateFrontFaceProps> = ({
     hideIssueDate,
     handleViewBackFace,
     showDetailsBtn = false,
+    formattedDisplayType,
 }) => {
     const {
         title = '',
@@ -145,6 +147,11 @@ export const CertificateFrontFace: React.FC<CertificateFrontFaceProps> = ({
                     <h1 className="text-grayscale-900 text-center text-[20px] font-jacques">
                         {title}
                     </h1>
+                    <div
+                        className={`${textLightColor} uppercase text-[14px] font-notoSans font-[600]`}
+                    >
+                        {formattedDisplayType}
+                    </div>
                 </div>
 
                 {customBodyCardComponent}
@@ -206,7 +213,9 @@ export const CertificateFrontFace: React.FC<CertificateFrontFaceProps> = ({
 
                     <VerifierStateBadgeAndText verifierState={verifierState} />
                 </div>
-                <div className={`${textLightColor} uppercase text-[14px] font-notoSans font-[600]`}>
+                <div
+                    className={`${textLightColor} uppercase text-[14px] font-notoSans font-[600]`}
+                >
                     {categoryTitle}
                 </div>
             </div>
@@ -222,7 +231,7 @@ export const CertificateFrontFace: React.FC<CertificateFrontFaceProps> = ({
 
             {!isSelfVerified && (
                 <CertificateProfileImageDisplay
-                    imageUrl={issuerImage}
+                    imageUrl={issuerImage || credential?.issuer?.image}
                     imageComponent={issuerImageComponent}
                     className={`w-[calc(100%-26px)] absolute bottom-0 flex justify-center items-center ${textDarkColor}`}
                     isIssuer