+++ services/learn-card-network/brain-service/src/routes/boosts.ts
@@ -111,12 +111,17 @@ export const boostsRouter = t.router({
                 profileId: z.string(),
                 uri: z.string(),
                 credential: VCValidator.or(JWEValidator),
+                options: z
+                    .object({
+                        skipNotification: z.boolean().default(false).optional(),
+                    })
+                    .optional(),
             })
         )
         .output(z.string())
         .mutation(async ({ ctx, input }) => {
             const { profile } = ctx.user;
-            const { profileId, credential, uri } = input;
+            const { profileId, credential, uri, options } = input;
 
             if (process.env.NODE_ENV !== 'test') {
                 console.log('🚀 BEGIN - Send Boost', JSON.stringify(input));
@@ -150,14 +155,17 @@ export const boostsRouter = t.router({
                 });
             }
 
-            return sendBoost(
-                profile,
-                targetProfile,
+            let skipNotification = profile.profileId === targetProfile.profileId;
+            if (options?.skipNotification) skipNotification = options?.skipNotification;
+
+            return sendBoost({
+                from: profile,
+                to: targetProfile,
                 boost,
                 credential,
-                ctx.domain,
-                profile.profileId === targetProfile.profileId
-            );
+                domain: ctx.domain,
+                skipNotification,
+            });
         }),
 
     createBoost: profileRoute